add: specific theme selection buttons
I would like to add that it is super early in the morning for me and I'm
tired so I'm not going to try spending my time adding the functionality
to add the "currently selected theme" outline to the buttons when using
other switch methods like the "random theme" button.

Instead those buttons will remove the outlines to prevent the issue of
the buttons outlining the wrong corresponding theme button.