style: remove comments
I plan on writing out documentation on why and how my site works, but
those comments are relics of the beginnings of my CSS file. So detailed
explanation on why my `:root` element contains the same valuables for
each variable as the dark theme can be explained later.

Oh, and a comment I forgot to remove that got rid of the
`outline: none;` declaration on buttons, as it was causing accessibility
issues at the time.