fix: accessibility issue with colored buttons
Due to the event listener in our Javascript targeting the element
surrounding our buttons, the browser thought we needed to add semantic
HTML to that element because it thought we were trying to make that a
button.

I went ahead and added the `role="none"` attribute to the element, which
made the error go away. I don't, however, use assistive technologies, so
I have not yet tested if this will actually make accessibility worse.