refactor: merge update theming functions
The function to swap the button that was currently representing the
"active" theme is only ever called in the `updateTheme` function, so
I threw all of the code from it into that function instead.

I also made the if else statement return instead of logging whenevr you
click something that isn't button (i.e., our button's parent div).