fix: overflow causing scroll & font size issues
Some flexbox elements on a given page would not show up at all if
their size eclipsed that of their parent containers.

Achieving a larger size than a parent container could happen in two
ways: Setting the font size to be large enough to cause overflow on
smaller devices, or to have content overflow in general.