refactor: make theme management less hard-coded
I originally wanted this project to be fairly easy to apply outside of
this little site of mine, so this was only inevitable. Hopefully this
makes things a lot easier in the long-run.

I will probably change the naming around "options" since I see it as a
better idea to call it something more like "configuration".