remove: CSS property order linting package
Frankly it was neat, but for what it offered, the size of the node
modules was ridiculous. It really doesn't matter as this is just a small
project that doesn't really have other contributors anyway.

If the project ever finds itself dealing with code inconsistency as a
result of my forgetfulness or more people contributing I will consider
adding it back in, but for now this is being removed to avoid
overcomplicating things on a personal project.