fix: to use dvh instad of vh for body height
This is because of a behavior on mobile browsers where the viewport size
would shift because of the browser UI. Because phones allow this UI to
collapse, the viewport size can change.

This change in the size of the viewport made for weird behavior where
users could scroll when the browser UI wasn't collapsed (which is the
default behavior, as our site doesn't have anywhere to scroll to).

Now users on mobile should be able to view the page without being able
to scroll, as they wouldn't have any reason to anyway.